Imp move OneHotEncoder to preprocessing/data.py
