+++ sklearn/preprocessing/__init__.py
@@ -7,16 +7,17 @@ from .data import Binarizer
 from .data import KernelCenterer
 from .data import MinMaxScaler
 from .data import Normalizer
+from .data import Scaler
 from .data import StandardScaler
 from .data import add_dummy_feature
 from .data import binarize
 from .data import normalize
 from .data import scale
+from .data import OneHotEncoder
 
 from .label import label_binarize
 from .label import LabelBinarizer
 from .label import LabelEncoder
-from .label import OneHotEncoder
 
 from .imputation import Imputer
 
@@ -29,6 +30,7 @@ __all__ = [
     'MinMaxScaler',
     'Normalizer',
     'OneHotEncoder',
+    'Scaler',
     'StandardScaler',
     'add_dummy_feature',
     'binarize',
